requirements.txt
python-telegram-bot==20.4
requests==2.31.0
aiohttp==3.8.4
dnspython==2.4.2
python-dotenv==1.0.0
